requirements.txt
# ─────────────────────────────────────────────
#  Histopathology CNN Classifier — Requirements
#  Python 3.10+ önerilir
# ─────────────────────────────────────────────

# Core Deep Learning
torch==2.2.2
torchvision==0.17.2

# Data & Preprocessing
numpy==1.26.4
Pillow==10.3.0
scikit-learn==1.4.2
scipy==1.13.0
opencv-python==4.9.0.80

# Stain Normalization
staintools==2.1.2          # Macenko / Reinhard normalization

# Training Utilities
tqdm==4.66.4               # progress bar
matplotlib==3.9.0          # loss/accuracy plots
seaborn==0.13.2            # confusion matrix heatmap

# Experiment Tracking (opsiyonel ama önerilir)
tensorboard==2.17.0        # eğitim grafiklerini izle

# Hyperparameter Optimization
scikit-optimize==0.10.1    # Bayesian optimization (BayesSearchCV)

# Model Export / Inference
onnx==1.16.0               # modeli ONNX'e export et (FastAPI için)
onnxruntime==1.18.0        # ONNX inference (GPU'suz sunucularda hızlı)

# Notebook (Colab / Kaggle dışında local çalışıyorsan)
ipykernel==6.29.4
ipywidgets==8.1.3
